requirements.txt
numpy>=1.23
pandas>=1.5
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
scikit-learn>=1.2
pyyaml>=6.0
rouge-score>=0.1.2
spacy>=3.5.0
nltk>=3.8.0
